==> notebooks/data_in_brief/1_construct_SCE_HIVE.Rmd <==
---
title: "Construct SCE for data in brief"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(scater)
library(scuttle)
library(Matrix)
library(here)
library(tidyverse)
```

# Construct single cell experiment object for data in brief

## Aim

The samples consist of 2 batches with different cell numbers fixed and processed with each device.

## Honeycomb HIVE

The data is not in a nice foraat that can directly be read into a SCE so I do it the slow way with the count matrix that is emitted by beenet.

Note the top 10,000 cell barcodes were written to file.

```{r read_sce_15k}
fiften <- read_tsv(here::here(
  "data/HIVE/AGI_15_20220628142630_TCM.tsv.gz"
  ))
fiften <- as.data.frame(fiften)
# First column is gene symbol
row.names(fiften) <- fiften[,1]
# Remove the row name
fiften <- as.matrix(fiften[,(2:10000)])
fiften <- Matrix(data = fiften, sparse = TRUE)

fiften <- SingleCellExperiment(assays=list(counts=fiften))
fiften$Fix_num <- 15000
```

Convert the other sample

```{r read_sce_30k}
thirty <- read_tsv(here::here(
  "data/HIVE/AGI_30_20220628140744_TCM.tsv.gz"
  ))
thirty <- as.data.frame(thirty)
row.names(thirty) <- thirty[,1]
thirty <- as.matrix(thirty[,(2:10000)])
thirty <- Matrix(data = thirty, sparse = TRUE)

thirty <- SingleCellExperiment(assays=list(counts=thirty))
thirty$Fix_num <- 30000
```

Combined the objects

```{r}
keep_genes <- intersect(row.names(fiften),
                        row.names(thirty))

sce <- cbind(fiften[keep_genes,], 
             thirty[keep_genes,])
```

## Save the basic SCE

```{r save_base}
sce_orig <- sce
saveRDS(sce, here::here(
  "data/SCEs/hive_basic_sce.rds"
  ))
```

## Add metadata details

Add in the metrics about fixation time, number of cells fixed etc.

Fixation time:

* Samples fixed on 11 April 2022
* Samples processed on 5 May 2022
* Fixation time is 3 weeks

```{r}
sce$Storage_time_weeks <- 3
sce$RBC_debris <- FALSE
sce$Donor <- "AS24"
```

## Compute QC metrics

Add per cell QC adds the QC info to the colData of the sce.

```{r}
is.mito <- grepl("^MT", rownames(sce))
sce <- addPerCellQC(sce, subsets=list(
  Mito=is.mito))
```

Add experimental metdata

```{r}
metadata(sce)$Platform <- "Honeycomb HIVE v1"
```

### Save SCE with quality control

```{r}
saveRDS(sce, here::here(
        "data/SCEs/hive_qcMetrics_sce.rds"
        ))
```

Check the sce object and check metadata

```{r, eval=FALSE}
sce <- readRDS(here::here(
  "data/SCEs/hive_qcMetrics_sce.rds"
))

colData(sce)
```

metadata is correct.

## Session Info

```{r}
sessionInfo()
```
