==> notebooks/data_in_brief/4_rough_clustering.Rmd <==
---
title: "Rough clustering for data in brief"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      warning = FALSE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/clustering/',
                      warning=FALSE, message=FALSE)
library(scuttle)
library(scater)
library(scran)
library(knitr)
library(tidyverse)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# Rough clustering for data in brief

## Aim

To generate some of the figures for data in brief, do a rough PCA, clustering and UMAP

# Dimension reduction and visualisation

## Honeycomb HIVE v1

Read data

```{r read_sce_hive}
hive <- readRDS(here::here(
  "data/SCEs/hive_qcMetrics_sce.rds"
  ))
```

Have to remove low quality cells or downstream steps fail due to missing values

```{r qc_hive}
is.mito <- grepl("^MT-", rownames(hive))
qcstats <- perCellQCMetrics(hive, subsets=list(Mito=is.mito))
filtered <- quickPerCellQC(qcstats, percent_subsets="subsets_Mito_percent")
hive <- hive[, !filtered$discard]
```

Normalisation and feature selection

```{r norm_feature_hive}
# Normalization.
hive <- logNormCounts(hive)

# Feature selection.
dec <- modelGeneVar(hive)
hvg <- getTopHVGs(dec, prop=0.1)
```

Principle component analysis
```{r pca_cluster_hive}
# PCA.
set.seed(1234)
hive <- runPCA(hive, ncomponents=12, subset_row=hvg)

# Clustering.
library(bluster)
colLabels(hive) <- clusterCells(hive, use.dimred='PCA',
    BLUSPARAM=NNGraphParam(cluster.fun="louvain"))    
```

UMAP

```{r umap_hive}
hive <- runUMAP(hive, dimred = 'PCA')

#saveRDS(hive, "/vast/scratch/users/brown.d/data_in_brief/hive_umap.rds")
```

## Visulaise UMAP {.tabset}

### By cluster label

```{r plot_umap_hive}
hive$fixed <- as.factor(hive$fixed)
p1 <- plotUMAP(hive, colour_by="label",rasterise=TRUE) + 
  xlab("") + ylab("") +
  theme_Publication()

p1
```

### By library size

```{r plot_umap_sum_hive}
hive$LibSize <- log(hive$sum)
p2 <- plotUMAP(hive, colour_by="LibSize",rasterise=TRUE) + 
  xlab("") + ylab("") +
  guides(colour=guide_legend(title="Log library \nsize")) +
  theme_Publication()

p2$layers[[1]]$aes_params$size <- 0.1
p2 + guides(color = guide_legend(override.aes = list(size = 2, alpha=1)))
```

### By batch

The 2 batches are the different cell inputs into a distinct hive device.

```{r plot_umap_sample_hive}
p3 <- plotUMAP(hive, colour_by="fixed",rasterise=TRUE) +
  xlab("") + ylab("") +
  guides(colour=guide_legend(title="Batch")) +
  guides(color = guide_legend(override.aes = list(size = 2, alpha=1))) +
  theme_Publication()

p3$layers[[1]]$aes_params$alpha <- 0.1
p3$layers[[1]]$aes_params$size <- 0.1
p3
```

## 10x Genomics Flex

Read data

```{r read_sce_flex}
flex <- readRDS(here::here(
  "data/SCEs/flex_qcMetrics.sce.rds"
  ))
```

Have to remove low quality cells or downstream steps fail due to missing values

```{r qc_flex}
is.mito <- grepl("^MT-", rownames(flex))
qcstats <- perCellQCMetrics(flex, subsets=list(Mito=is.mito))
filtered <- quickPerCellQC(qcstats, percent_subsets="subsets_Mito_percent")
flex <- flex[, !filtered$discard]
```

Normalisation and feature selection

```{r norm_feature_flex}
# Normalization.
flex <- logNormCounts(flex)

# Feature selection.
dec <- modelGeneVar(flex)
hvg <- getTopHVGs(dec, prop=0.1)
```

Principle component analysis
```{r pca_cluster_flex}
# PCA.
set.seed(1234)
flex <- runPCA(flex, ncomponents=12, subset_row=hvg)

# Clustering.
colLabels(flex) <- clusterCells(flex, use.dimred='PCA',
    BLUSPARAM=NNGraphParam(cluster.fun="louvain"))    
```

UMAP

```{r umap_flex}
flex <- runUMAP(flex, dimred = 'PCA')

#saveRDS(flex, "/vast/scratch/users/brown.d/data_in_brief/flex_umap.rds")
#flex <- readRDS("/vast/scratch/users/brown.d/data_in_brief/flex_umap.rds")
```

## Visulaise UMAP {.tabset}

### By cluster label

```{r plot_umap_flex}
p1 <- plotUMAP(flex, colour_by="label",rasterise=TRUE) +
  xlab("") + ylab("") +
  theme_Publication()

p1
```

### By library size

```{r plot_umap_sum_flex}
flex$LibSize <- log(flex$sum)
p2 <- plotUMAP(flex, colour_by="LibSize", rasterise=TRUE) + 
  xlab("") + ylab("") +
  guides(colour=guide_legend(title="Log library \nsize")) +
  theme_Publication()

p2$layers[[1]]$aes_params$size <- 0.1
p2 + guides(color = guide_legend(override.aes = list(size = 2, alpha=1)))
```

### By batch

The 2 batches are the different cell inputs into a distinct flex device.

```{r plot_umap_sample_flex}
p3 <- plotUMAP(flex, colour_by="Batch", rasterise=TRUE) +
  xlab("") + ylab("") +
  guides(color = guide_legend(override.aes = list(size = 2, alpha=1))) +
  theme_Publication()

p3$layers[[1]]$aes_params$alpha <- 0.1
p3$layers[[1]]$aes_params$size <- 0.1
p3
```


## Session Info

```{r}
sessionInfo()
```
